chore: all demos
