+++ src/client/theme-api/context/index.ts
@@ -1,7 +1,8 @@
 import type { PICKED_PKG_FIELDS } from '@/constants';
 import type { AtomComponentAsset } from 'dumi-assets-types';
 import { createContext, useContext, type ComponentType } from 'react';
-import type { ILocalesConfig, IPreviewerProps, IThemeConfig } from './types';
+import type { ILocalesConfig, IPreviewerProps, IThemeConfig } from '../types';
+import use from './use';
 
 export type DemoInfo = {
   component: ComponentType;
@@ -48,3 +49,16 @@ export const SiteContext = createContext<ISiteContext>({
 export const useSiteData = () => {
   return useContext(SiteContext);
 };
+
+const cache = new Map<string, Promise<DemoInfo | null>>();
+
+// Async load demo data
+export function useDemoData(demoId: string) {
+  const { getDemoById } = useSiteData();
+
+  if (!cache.has(demoId)) {
+    cache.set(demoId, getDemoById(demoId));
+  }
+
+  return use(cache.get(demoId)!);
+}
